requirements.txt
requests
Pillow
RPI.GPIO
spidev
